backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-chroma
langchain-groq
langchain-huggingface
langgraph
pypdf
chromadb
python-dotenv
tiktoken
sentence-transformers